remaining

+++ functions.py
@@ -1,6 +1,6 @@
 # FETCHALL RETORNA UMA LISTA DE TUPLAS, QUE SAO AS LINHAS DO DB, MESMO QUE SO HAJA 1 LINHA SELECIONADA. LOGO, PARA ACESSAR OS ELEMENTOS, PRECISA ESPECIFICAR DOIS INDICES (DBLIST[0][1] = NOME DO PRODUTO; ln.179) OU, NO CASO DE HAVER 1 LINHA SO, TORNAR ESSA LISTA UMA TUPLA DESSA LINHA (dblist = dblist[0]) E ENTAO ACESSAR OS ELEMENTOS COM UM UNICO INDICE (dblist[x]; ln.183)
 
-import sqlite3, os, datetime, smtplib
+import sqlite3, os, datetime, smtplib, mysql.connector
 
 
 def arq_promocao(modo):
@@ -39,15 +39,34 @@ class banco_de_dados():
     def __init__(self):
         pass
 
-    def __conectardb(self):
-        return sqlite3.connect("%s\\db.db" % (os.getcwd()))
+    def __conectardbuser():
+        db = mysql.connector.connect(
+            host="localhost",
+            port="3307",
+            user='clientes',
+            passwd="@L0j4M4r0mb4#Cl13nt3",
+            database="lojadb"
+        )
+        return db
+
+
+    def __conectardbadmin():
+        db = mysql.connector.connect(
+                host="localhost",
+                port="3307",
+                user='admin',
+                passwd="@L0j4M4r0mb4",
+                database="lojadb"
+        )
+        return db
+
 
     def cadastrar_usuario(self, dados):
 
-        db = self.__conectardb()
+        db = self.__conectardbadmin()
         cur = db.cursor()
         try:
-            cur.execute("INSERT INTO usuarios VALUES (?, ?, ?, ?)", (dados))
+            cur.execute("INSERT INTO usuarios (email, nome, senha) VALUES ('%s', '%s', '%s')" % (dados[2],dados[1],dados[3]))
             db.commit()
             db.close()
             return True
@@ -57,7 +76,7 @@ class banco_de_dados():
 
     def buscar_usuario(self, email):
 
-        db = self.__conectardb()
+        db = self.__conectardbuser()
         cur = db.cursor()
         try:
             cur.execute("SELECT nome, senha FROM usuarios where email = '%s'" %(email))
@@ -71,12 +90,12 @@ class banco_de_dados():
     def add_item(self, codigo, nome, preco, categoria, arq_imagem, descricao):
 
 
-        db = self.__conectardb()
-        cur = db.cursor()
         try:
-            dataAdd = int(data_atual())
-            cur.execute("""INSERT INTO itens values (?,?,?,?,?,?,?)
-            """, (codigo, nome, preco, categoria, arq_imagem, dataAdd, None))
+            print(codigo, nome, preco, categoria, arq_imagem, descricao)
+            db = self.__conectardbadmin()
+            cur = db.cursor()
+            dataAdd = data_atual()
+            cur.execute("INSERT INTO itens (codigo, nome, preco, categoria, imagem, data) values ('%s', '%s', '%s', '%s', '%s', '%s')" % (codigo, nome, preco, categoria, arq_imagem, dataAdd))
             db.commit()
             db.close()
             arq = open(str(os.getcwd())+"\\descricoes\\%s.txt" % (codigo), 'w')
@@ -89,7 +108,7 @@ class banco_de_dados():
 
     def del_item(self, codigo):
 
-        db = self.__conectardb()
+        db = self.__conectardbadmin()
         cur = db.cursor()
         try:
             os.remove(str(os.getcwd())+"\\descricoes\\%s.txt" % (codigo))
@@ -97,25 +116,26 @@ class banco_de_dados():
         except FileNotFoundError: 
             status = "Arquivo da descrição não encontrado."
 
-        try:
+        cur.execute("SELECT codigo FROM itens WHERE codigo = '%s'" % (codigo))
+        existe = cur.fetchall()
+        
+        if len(existe) == 0: 
+            status = status+" Codigo não existe"
+            return status
+        else:
             cur.execute("""DELETE FROM itens WHERE codigo = '%s'""" % (codigo))
             db.commit()
             db.close()
             status = status+" Item removido"
             return status
-        except:
-            db.close()
-            status = status+" Codigo não existe"
-            return status
-
+            
     def att_preco(self, codigo, preco_novo):
 
-        db = self.__conectardb()
+        db = self.__conectardbadmin()
         cur = db.cursor()
         try:
-            sql = "UPDATE itens SET preco = ? WHERE codigo = ?"
-            values = (preco_novo, codigo)
-            cur.execute(sql, (preco_novo, codigo))
+            sql = "UPDATE itens SET preco = '%s' WHERE codigo = '%s'" % (preco_novo, codigo)
+            cur.execute(sql)
             db.commit()
             db.close()
             return True
@@ -125,33 +145,32 @@ class banco_de_dados():
 
     def listar_categoria(self, categoria):
 
-        db = self.__conectardb()
+        db = self.__conectardbuser()
         cur = db.cursor()
-        cur.execute("""SELECT * FROM itens where categoria = '%s'
-        """ % (categoria))
+        cur.execute("SELECT * FROM itens where categoria = '%s'" % (categoria))
         itens = cur.fetchall()
         db.close()
         return itens
 
     def att_senha(self, email, senha_nova):
 
-        db = self.__conectardb()
+        db = self.__conectardbadmin()
         cur = db.cursor()
         cur.execute("""UPDATE usuarios
-                            SET senha = '%s'
-                            WHERE email = '%s'
-        """ % (senha, email))
+                            SET senha = '?'
+                            WHERE email = '?'
+        """, (senha, email))
         db.commite()
         db.close()
         return True
 
     def add_promocao(self, codigo_item, preco_novo):
 
-        db = self.__conectardb()
+        db = self.__conectardbuser()
         cur = db.cursor()
 
         try:
-            cur.execute("""SELECT * FROM itens WHERE codigo = %i""" % (int(codigo_item)))
+            cur.execute("""SELECT * FROM itens WHERE codigo = %i""" % (int(codigo_item),))
             item = cur.fetchall()
             preco_antigo = item[0][2]
         except:
@@ -193,13 +212,9 @@ class banco_de_dados():
     def buscar_itens_novos(self):
 
         i = int(data_atual())-7
-        db = self.__conectardb()
+        db = self.__conectardbuser()
         cur = db.cursor()
-        cur.execute("""SELECT codigo, nome, preco, imagem
-        FROM itens
-        WHERE data > %i
-        ORDER BY data asc;
-        """ % (i))
+        cur.execute("""SELECT codigo, nome, preco, imagem FROM itens WHERE data > %i ORDER BY data asc """ % (i))
         itens = cur.fetchall()
         db.close()
         # itens = [codigo, nome, preco, imagem]
@@ -213,7 +228,7 @@ class banco_de_dados():
         except:
             arq = arq_promocao("w")
         arq.close()
-        db = self.__conectardb()
+        db = self.__conectardbuser()
         cur = db.cursor()
 
         for line in range(len(itens)):
@@ -228,24 +243,24 @@ class banco_de_dados():
         return itens
 
     def buscar_item(self, codigo):
-        db = self.__conectardb()
+        db = self.__conectardbuser()
         cur = db.cursor()
         
-        cur.execute("SELECT * FROM itens WHERE codigo = ?", codigo)
+        cur.execute("SELECT * FROM itens WHERE codigo = '%s'" % (codigo))
         infos = cur.fetchall()
         if infos == []:
             return False
         else:
-            infos = [infos[0][0],infos[0][1],infos[0][2],infos[0][3],infos[0][4],infos[0][5],infos[0][6]]
+            infos = [infos[0][0],infos[0][1],infos[0][2],infos[0][3],infos[0][4],infos[0][5]]
             db.close()
             descricao = arq_descricao(codigo[0])
             infos.append(descricao)
             return infos
 
     def contar_itens(self, categ):
-        db = self.__conectardb()
+        db = self.__conectardbuser()
         cur = db.cursor()
-        cur.execute("""SELECT nome FROM itens where categoria = '%s'""" % (categ))
+        cur.execute("SELECT nome FROM itens where categoria = '%s'" % (categ))
         dblist = cur.fetchall()
         dblist = len(dblist)
         db.close()
@@ -258,25 +273,28 @@ class banco_de_dados():
         return int(codigo_produto)
 
     def enviar_duvida(self, codigo, nome, email, duvida):
-        db = self.__conectardb()
+        db = self.__conectardbadmin()
         cur = db.cursor()
-        cur.execute("""
-                    SELECT codigo from duvidas WHERE email = (?)
-        """, (email,))
-        duvidas_anteriores = cur.fetchall()
-        for x in range(len(duvidas_anteriores)):
-            duvidas_anteriores[x] = duvidas_anteriores[x][0]
-        if int(codigo) in duvidas_anteriores:
-            return False
-        else:
-            cur.execute("""INSERT INTO duvidas values (?,?,?,?)""", (codigo, nome, email, duvida))
+        # cur.execute("""
+        #             SELECT codigo from duvidas WHERE email = '?'
+        # """, (email))
+        # duvidas_anteriores = cur.fetchall()
+        # for x in range(len(duvidas_anteriores)):
+        #     duvidas_anteriores[x] = duvidas_anteriores[x][0]
+        # if int(codigo) in duvidas_anteriores:
+        #     return False
+        # else:
+        try:
+            cur.execute("""INSERT INTO duvidas (codigo, nome, email, duvida) values ('%s','%s','%s','%s')""" % (codigo, nome, email, duvida))
             db.commit()
             db.close()
             return "Duvida enviada. Aguarde um email de resposta."
+        except mysql.connector.errors.IntegrityError:
+            return False
 
     def listar_duvidas(self):
 
-        db = self.__conectardb()
+        db = self.__conectardbuser()
         cur = db.cursor()
         cur.execute("""
                     SELECT * FROM duvidas
@@ -286,8 +304,8 @@ class banco_de_dados():
     
     def deletar_duvida(self, nome, codigo):
 
-        db = self.__conectardb()
+        db = self.__conectardbadmin()
         cur = db.cursor()
-        cur.execute("DELETE FROM duvidas WHERE codigo = (?) AND nome = (?) ", (codigo, nome))
+        cur.execute("DELETE FROM duvidas WHERE codigo = %i AND nome = %s " %(codigo, nome))
         db.commit()
         db.close()
